# src/core/core/static/openapi3_0.yaml
openapi: 3.0.3
info:
  title: Taranis NG
  description: |-
    The Taranis NG API provides endpoints for the Taranis NG is an OSINT gathering and analysis tool.
  license:
    name: EUPL-1.2
  version: 0.1.8
...
servers:
  - url: /api/v1
    description: default
  - url: https://dev.taranis.cyberrange.rocks/api/v1
    description: dev
  - url: https://taranis.cyberrange.rocks/api/v1
    description: main
...
security:
  - API_key: []
  - user_auth: []
paths:
...
  /config/reload-enum-dictionaries/{dictionary_type}:
    parameters:
      - name: dictionary_type
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [config]
      summary: reload dictionaries
      # TODO: this should be an HTTP POST/PUT request
      responses:
        '200':
          description: successfully reloaded dictionaries
          content:
            application/json:
              schema:
                type: string
                example: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
  /config/attributes:
    get:
      tags: [config]
      summary: get attributes
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: attributes
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/attribute'
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
    post:
      tags: [config]
      summary: add attribute
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/attribute.new'
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
  /config/attributes/{attribute_id}:
    parameters:
      - name: attribute_id
        in: path
        required: true
        schema:
          type: integer
    put:
      tags: [config]
      summary: update attribute
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/attribute.new'
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
    delete:
      tags: [config]
      summary: delete attribute
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
  /config/attributes/{attribute_id}/enums:
    parameters:
      - name: attribute_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [config]
      summary: get attribute enum
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: offset
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
    post:
      tags: [config]
      summary: add attribute enums
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                delete_existing:
                  type: boolean
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/attribute_enum.new'
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
  /config/attributes/{attribute_id}/enums/{enum_id}:
    parameters:
      - name: attribute_id
        in: path
        required: true
        schema:
          type: integer
      - name: enum_id
        in: path
        required: true
        schema:
          type: integer
    put:
      tags: [config]
      summary: update attribute enum
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/attribute_enum.new'
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
    delete:
      tags: [config]
      summary: delete attribute enum
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
  /config/report-item-types:
    get:
      tags: [config]
      summary: get report item types
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/report_item_type.presentation'
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
    post:
      tags: [config]
      summary: add report item type
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/report_item_type.new'
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
  /config/report-item-types/{type_id}:
    parameters:
      - name: type_id
        in: path
        required: true
        schema:
          type: integer
    put:
      tags: [config]
      summary: update report item type
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/report_item_type.new'
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
    delete:
      tags: [config]
      summary: delete report item type
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
  /config/product-types:
    get:
      tags: [config]
      summary: get all product types
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/product_type.presentation'
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
    post:
      tags: [config]
      summary: add product type
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/product_type.new'
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
  /config/product-types/{type_id}:
    parameters:
      - name: type_id
        in: path
        required: true
        schema:
          type: integer
    put:
      tags: [config]
      summary: update product type
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/product_type.new'
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
    delete:
      tags: [config]
      summary: delete product type
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
  /config/permissions:
    get:
      tags: [config]
      summary: get all permissions
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/permission'
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
...
  /config/roles:
    get:
      tags: [config]
      summary: get all roles
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/role'
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
    post:
      tags: [config]
      summary: add role
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/role'
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
  /config/roles/{role_id}:
    parameters:
      - name: role_id
        in: path
        required: true
        schema:
          type: integer
    put:
      tags: [config]
      summary: update role
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/role'
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
    delete:
      tags: [config]
      summary: delete role
      responses:
        '200':
          description: success
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
...
  /config/export-osint-sources:
    post:
      tags: [config]
      summary: export OSINT sources
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                selection:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/osint_source.export.root'
        '401':
          description: authorization error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resp.auth.error'
        '404':
          $ref: "#/components/schemas/notFound"
        '405':
          description: method not allowed
          content:
            'application/json': { }
            'application/xml': { }
...
components:
  schemas:
...
    notFound:
      description: "Not found response"
...
    resp.auth.error:
      type: object
      properties:
        error:
          type: string
...
    report_item_type.new:
      allOf:
        - $ref: '#/components/schemas/report_item_type.base'
        - type: object
          properties:
            attribute_groups:
              type: array
              items:
                $ref: '#/components/schemas/attribute_group.new'
...
    report_item_type.presentation:
      allOf:
        - $ref: '#/components/schemas/report_item_type'
        - $ref: '#/components/schemas/presentation'
...
    attribute_enum.new:
      allOf:
        - $ref: '#/components/schemas/attribute_enum'
    attribute.new:
      allOf:
        - $ref: '#/components/schemas/attribute.base'
        - type: object
          properties:
            attribute_enums:
              type: array
              items:
                $ref: '#/components/schemas/attribute_enum.new'
...
    attribute:
      allOf:
        - $ref: '#/components/schemas/attribute.base'
        - type: object
          properties:
            attribute_enums:
              type: array
              items:
                $ref: '#/components/schemas/attribute_enum'
...
  securitySchemes:
    API_key:
      type: http
      scheme: bearer
      bearerFormat: APIKEY
    user_auth:
      type: http
      schema: bearer
      bearerFormat: JWT